move to utils, add tests
